requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
absl-py=0.11.0=py37h89c1867_0
aiohttp=3.7.3=py37h4abf009_0
appdirs=1.4.4=pyh9f0ad1d_0
async-timeout=3.0.1=py_1000
attrs=20.3.0=pyhd3deb0d_0
audioread=2.1.9=py37h89c1867_0
backcall=0.2.0=py_0
blas=1.0=mkl
blinker=1.4=py_1
boto=2.49.0=py37_0
boto3=1.9.66=py37_0
botocore=1.12.67=py37_0
bravado=11.0.2=pyh9f0ad1d_0
bravado-core=5.17.0=pyh9f0ad1d_0
brotlipy=0.7.0=py37hb5d75c8_1001
bz2file=0.98=pypi_0
bzip2=1.0.8=h7f98852_4
c-ares=1.17.1=h36c2ea0_0
ca-certificates=2020.12.8=h06a4308_0
cachetools=4.1.1=py_0
cairo=1.14.12=h8948797_3
certifi=2020.12.5=py37h06a4308_0
cffi=1.14.4=py37h261ae71_0
chardet=3.0.4=py37he5f6b98_1008
click=7.1.2=pyh9f0ad1d_0
cryptography=3.2.1=py37hc72a4ac_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=py_2
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
docutils=0.16=py37_1
essentia=2.1b6.dev234=pypi_0
expat=2.2.10=he6710b0_2
ffmpeg=4.0=hcdf2ecd_0
filelock=3.0.12=pypi_0
fire=0.3.1=pypi_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.10.4=h5ab3b9f_0
fsspec=0.8.4=py_0
future=0.18.2=py37h89c1867_2
gensim=3.8.0=py37h962f231_0
gettext=0.19.8.1=h0b5b191_1005
gitdb=4.0.5=py_0
gitpython=3.1.11=py_0
glib=2.66.1=h92f7085_0
glob2=0.7=py_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-api-core=1.22.2=py37_0
google-auth=1.23.0=pyhd8ed1ab_0
google-auth-oauthlib=0.4.1=py_2
google-cloud-core=1.4.4=pyhd3eb1b0_0
google-cloud-storage=1.31.0=py_0
google-crc32c=1.0.0=py37h7b6447c_0
google-resumable-media=1.1.0=py_1
googleapis-common-protos=1.52.0=py37_0
graphite2=1.3.14=h23475e2_0
grpcio=1.33.2=py37haffed2e_2
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
idna=2.10=pyh9f0ad1d_0
importlib-metadata=3.1.1=pyhd8ed1ab_0
importlib_metadata=3.1.1=hd8ed1ab_0
intel-openmp=2020.2=254
ipython=7.19.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jasper=2.0.14=h07fcdf6_1
jedi=0.17.2=py37h06a4308_1
jmespath=0.10.0=py_0
joblib=0.17.0=py_0
jpeg=9b=h024ee3a_2
jsonref=0.2=py_0
jsonschema=3.2.0=py_2
kiwisolver=1.3.1=py37hc928c03_0
lame=3.100=h14c3975_1001
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libcrc32c=1.1.1=he6710b0_2
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libflac=1.3.3=h9c3ff4c_1
libgcc-ng=9.3.0=h5dbcf3e_17
libgfortran-ng=7.5.0=hae1eefd_17
libgfortran4=7.5.0=hae1eefd_17
libglu=9.0.0=hf484d3e_1
libgomp=9.3.0=h5dbcf3e_17
libiconv=1.16=h516909a_0
libllvm10=10.0.1=he513fc3_3
libogg=1.3.4=h7f98852_0
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7f98852_1
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=h8b12597_0
librosa=0.8.0=pyh9f0ad1d_0
libsndfile=1.0.30=h9c3ff4c_1
libstdcxx-ng=9.3.0=h2ae2ef3_17
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libvorbis=1.3.7=he1b5a44_0
libvpx=1.7.0=h439df22_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
llvmlite=0.35.0=py37h9d7f4d0_0
lz4-c=1.9.2=heb0550a_3
markdown=3.3.3=pyh9f0ad1d_0
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py37h817c723_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
monotonic=1.5=py_0
msgpack-python=1.0.0=py37hc928c03_2
multidict=4.7.5=py37h8f50634_2
ncurses=6.2=he6710b0_1
neptune-client=0.4.128=pyhd3deb0d_0
nettle=3.6=he412f7d_0
ninja=1.10.2=py37hff7bd54_0
nnaudio=0.2.0=pypi_0
numba=0.52.0=py37hdc94413_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
oauthlib=3.0.1=py_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openh264=2.1.1=h8b12597_0
openssl=1.1.1i=h27cfd23_0
packaging=20.7=pyhd3deb0d_0
pandas=1.1.3=py37he6710b0_0
parso=0.7.0=py_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.0.1=py37he98fc37_0
pip=20.3.1=py37h06a4308_0
pixman=0.40.0=h7b6447c_0
pooch=1.3.0=pyhd8ed1ab_0
prompt-toolkit=3.0.8=py_0
protobuf=3.13.0.1=py37he6710b0_1
ptyprocess=0.6.0=pyhd3eb1b0_2
py-opencv=3.4.2=py37hb342d67_1
py3nvml=0.2.5=py37hc8dfbb8_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.3=pyhd3eb1b0_0
pyjwt=1.7.1=py_0
pyopenssl=20.0.0=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.17.3=py37h4abf009_1
pysocks=1.7.1=py37he5f6b98_2
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.7.9=h7579374_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.7.0=py3.7_cuda10.1.243_cudnn7.6.3_0
pytorch-lightning=1.0.8=pyhd8ed1ab_0
pytorch-metric-learning=0.9.94=pyh39e3cac_0
pytz=2020.4=pyhd8ed1ab_0
pyyaml=5.3.1=py37hb5d75c8_1
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
regex=2020.11.13=pypi_0
requests=2.25.0=pyhd3deb0d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
resampy=0.2.2=py_0
rsa=4.6=pyh9f0ad1d_0
s3transfer=0.1.13=py37_0
sacremoses=0.0.43=pypi_0
scikit-learn=0.23.2=py37h0573a6f_0
scipy=1.5.2=py37h0b6359f_0
seaborn=0.11.0=py_0
setuptools=51.0.0=py37h06a4308_2
simplejson=3.17.2=py37h8f50634_1
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py37h06a4308_0
smart_open=2.0.0=py_0
smmap=3.0.4=pyh9f0ad1d_0
sqlite=3.33.0=h62c20be_0
swagger-spec-validator=2.7.3=pyh9f0ad1d_0
tbb=2020.2=hc9558a2_0
tensorboard=2.4.0=pyhd8ed1ab_0
tensorboard-plugin-wit=1.7.0=pyh9f0ad1d_0
termcolor=1.1.0=pypi_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tokenizers=0.9.4=pypi_0
torchaudio=0.7.0=py37
torchvision=0.8.1=py37_cu101
tornado=6.1=py37h4abf009_0
tqdm=4.54.1=pyhd8ed1ab_0
traitlets=5.0.5=py_0
transformers=4.0.0=pypi_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
umap-learn=0.4.6=py37hc8dfbb8_0
urllib3=1.24.3=py37_0
wcwidth=0.2.5=py_0
websocket-client=0.57.0=py37hc8dfbb8_3
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.1=pyhd3eb1b0_0
x264=1!152.20180806=h14c3975_0
xmltodict=0.12.0=py_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
yarl=1.6.3=py37h4abf009_0
zipp=3.4.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
